code/python/tests-workshop/requirements.txt
aws-cdk-lib==2.0.0-rc.29
constructs>=10.0.0,<11.0.0
cdk-dynamo-table-viewer==0.1.139
